Handle incomplete image size in HarbourImageProvider
If only one dimension is specified, the image gets loaded twice,
second time scaled to the specified dimension. Reloading is necessary
for SVGs (and probably only for SVGs), to make sure that they get
rendered according to the specified pixel resolution. That's not a
big deal though, since this provider is mostly being used to load
SVGs.