Added more Base32 encoding options
Now it's possible not only to choose upper/lower case but also to opt out
of generating the padding. In some cases (e.g. for generating otpauth URLs)
padding it not needed.

This slightly changes the API. The HarbourBase32::toBase32 call without any
options now generates uppercase Base32 rather than the lower case - because
RFC 4648 specifies the Base32 alphabet in upper case (although kind of hints
that it's actually case-insensitive)